requirements/dev.txt
alembic==1.0.0
atomicwrites==1.1.5
attrs==18.1.0
bcrypt==3.1.4
beautifulsoup4==4.6.1
blinker==1.4
cffi==1.11.5
click==6.7
dominate==2.3.1
factory-boy==2.11.1
Faker==0.8.17
flake8==3.5.0
flake8-blind-except==0.1.1
flake8-debugger==3.1.0
flake8-docstrings==1.3.0
flake8-isort==2.5
flake8-polyfill==1.0.2
flake8-quotes==1.0.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.4.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.2.0
Flask-Principal==0.4.0
Flask-Redis==0.3.0
Flask-SQLAlchemy==2.3.2
Flask-SSE==0.2.1
Flask-Webpack==0.1.0
Flask-WTF==0.14.2
gunicorn==19.9.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
#mysqlclient==1.3.13
pep8-naming==0.7.0
Pillow==5.2.0
pluggy==0.6.0
psutil==5.4.6
psycopg2==2.7.5
py==1.5.4
pycodestyle==2.3.1
pycparser==2.18
pydocstyle==2.1.1
pyflakes==1.6.0
pytest==3.6.3
python-dateutil==2.7.3
python-editor==1.0.3
redis==2.10.6
simplejson==3.16.0
six==1.11.0
snowballstemmer==1.2.1
SQLAlchemy==1.2.10
testfixtures==6.2.0
text-unidecode==1.2
visitor==0.1.3
waitress==1.1.0
WebOb==1.8.2
WebTest==2.0.30
Werkzeug==0.14.1
WTForms==2.2.1
